requirements.txt
asgiref==3.8.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
cryptography==43.0.1
dj-database-url==2.2.0
Django==5.1.1
django-allauth==65.0.2
django-heroku==0.3.1
gunicorn==23.0.0
idna==3.10
packaging==24.1
psycopg2==2.9.9
pycparser==2.22
PyJWT==2.9.0
requests==2.32.3
sqlparse==0.5.1
typing_extensions==4.12.2
urllib3==2.2.3
whitenoise==6.7.0
asgiref==3.8.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
cryptography==43.0.1
dj-database-url==2.2.0
Django==5.1.1
django-allauth==65.0.2
django-heroku==0.3.1
gunicorn==23.0.0
idna==3.10
packaging==24.1
pillow==10.4.0
psycopg2==2.9.9
pycparser==2.22
PyJWT==2.9.0
requests==2.32.3
sqlparse==0.5.1
typing_extensions==4.12.2
urllib3==2.2.3
whitenoise==6.7.0
asgiref==3.8.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
cryptography==43.0.1
dj-database-url==2.2.0
Django==5.1.1
django-allauth==65.0.2
django-crispy-forms==2.3
django-heroku==0.3.1
gunicorn==23.0.0
idna==3.10
packaging==24.1
pillow==10.4.0
psycopg2==2.9.9
pycparser==2.22
PyJWT==2.9.0
requests==2.32.3
sqlparse==0.5.1
typing_extensions==4.12.2
urllib3==2.2.3
whitenoise==6.7.0
asgiref==3.8.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
crispy-bootstrap4==2024.10
cryptography==43.0.1
dj-database-url==2.2.0
Django==5.1.1
django-allauth==65.0.2
django-crispy-forms==2.3
django-heroku==0.3.1
gunicorn==23.0.0
idna==3.10
packaging==24.1
pillow==10.4.0
psycopg2==2.9.9
pycparser==2.22
PyJWT==2.9.0
requests==2.32.3
sqlparse==0.5.1
typing_extensions==4.12.2
urllib3==2.2.3
whitenoise==6.7.0
asgiref==3.8.1
boto3==1.35.39
botocore==1.35.39
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
crispy-bootstrap4==2024.10
cryptography==43.0.1
dj-database-url==2.2.0
Django==5.1.1
django-allauth==65.0.2
django-crispy-forms==2.3
django-heroku==0.3.1
django-storages==1.14.4
gunicorn==23.0.0
idna==3.10
jmespath==1.0.1
packaging==24.1
pillow==10.4.0
psycopg2==2.9.9
pycparser==2.22
PyJWT==2.9.0
python-dateutil==2.9.0.post0
requests==2.32.3
s3transfer==0.10.3
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
urllib3==2.2.3
whitenoise==6.7.0
asgiref==3.8.1
boto3==1.35.39
botocore==1.35.39
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
crispy-bootstrap4==2024.10
cryptography==43.0.1
dj-database-url==2.2.0
Django==5.1.1
django-allauth==65.0.2
django-crispy-forms==2.3
django-heroku==0.3.1
django-storages==1.14.4
gunicorn==23.0.0
idna==3.10
jmespath==1.0.1
packaging==24.1
pillow==10.4.0
psycopg2==2.9.9
pycparser==2.22
PyJWT==2.9.0
python-dateutil==2.9.0.post0
requests==2.32.3
s3transfer==0.10.3
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
urllib3==2.2.3
whitenoise==6.7.0
